Add protection against possible null pointer and remove dead (unreachable) code (coverity)
git-svn-id: http://root.cern.ch/svn/root/trunk@32186 27541ba8-7e3a-0410-8455-c3a389f83636

+++ gui/gui/src/TGFileBrowser.cxx
@@ -547,8 +547,9 @@ void TGFileBrowser::Update()
          fListTree->GetPathnameFromItem(curr->GetParent(), path);
          if (strlen(path) > 1) {
             TString dirpath = FullPathName(curr->GetParent());
-            gSystem->GetPathInfo(dirpath.Data(), &id, &size, &flags, &modtime);
-            if (flags & 2) {
+            Int_t res = gSystem->GetPathInfo(dirpath.Data(), &id, &size, 
+                                             &flags, &modtime);
+            if ((res == 0) && (flags & 2)) {
                TString fullpath = FullPathName(curr);
                if (gSystem->AccessPathName(fullpath.Data()))
                   fListTree->DeleteItem(curr);
